+++ src/app/(main)/artist/_components/mainPage/Artist.tsx
