+++ src/app/(main)/layout.tsx
@@ -1,7 +1,8 @@
 import Header from "@/components/Header";
 import React, { PropsWithChildren } from "react";
-import NavigationBar from "./_components/NavigationBar";
+import NavigationBar from "@/components/NavigationBar";
 import Profile from "./_components/Profile";
+
 
 const MainLayout = ({ children }: PropsWithChildren) => {
   return (
