+++ src/app/(main)/page.tsx
@@ -1,9 +1,9 @@
 import LogoutButton from "@/components/LogoutButton";
 import React from "react";
-import Trending from "./_components/Trending";
-import Playlists from "./_components/Playlists";
+import Trending from "./playlists/_components/Trending";
+import Playlists from "./playlists/_components/Playlists";
 import Profile from "./_components/Profile";
-import Artist from "./_components/Artist";
+import Artist from "./artist/_components/mainPage/Artist";
 
 const MainPage = () => {
   return (
